+++ examples/src/main/java/io/opentdf/platform/DecryptCollectionExample.java
@@ -1,19 +1,12 @@
 package io.opentdf.platform;
 
 import io.opentdf.platform.sdk.Config;
-import io.opentdf.platform.sdk.NanoTDF;
 import io.opentdf.platform.sdk.SDK;
 import io.opentdf.platform.sdk.SDKBuilder;
 
-import java.io.ByteArrayInputStream;
 import java.io.FileInputStream;
-import java.io.FileOutputStream;
 import java.io.IOException;
-import java.net.URISyntaxException;
 import java.nio.ByteBuffer;
-import java.nio.charset.StandardCharsets;
-import java.security.NoSuchAlgorithmException;
-import java.util.concurrent.ExecutionException;
 
 public class DecryptCollectionExample {
     public static void main(String[] args) throws IOException {
