+++ sdk/src/main/java/io/opentdf/platform/sdk/ECKeyPair.java
@@ -1,6 +1,5 @@
-package io.opentdf.platform.sdk.nanotdf;
+package io.opentdf.platform.sdk;
 
-import io.opentdf.platform.sdk.KeyType;
 import org.bouncycastle.asn1.pkcs.PrivateKeyInfo;
 import org.bouncycastle.asn1.x509.SubjectPublicKeyInfo;
 import org.bouncycastle.cert.X509CertificateHolder;
