+++ sdk/src/main/java/io/opentdf/platform/sdk/NanoTDF.java
@@ -4,7 +4,6 @@ import com.connectrpc.ResponseMessageKt;
 import io.opentdf.platform.policy.kasregistry.ListKeyAccessServersRequest;
 import io.opentdf.platform.policy.kasregistry.ListKeyAccessServersResponse;
 import io.opentdf.platform.sdk.SDK.KasAllowlistException;
-import io.opentdf.platform.sdk.nanotdf.*;
 
 import java.io.IOException;
 import java.io.OutputStream;
