+++ sdk/src/test/java/io/opentdf/platform/sdk/ResourceLocatorTest.java
@@ -1,7 +1,8 @@
-package io.opentdf.platform.sdk.nanotdf;
+package io.opentdf.platform.sdk;
 
 import java.nio.ByteBuffer;
 import java.util.stream.Stream;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
